Hide todo-badger from itself
Ensure the regex we build doesn't match the string `'todo-badger"`. This string
needs to appear in the `.github/workflows/foo.yml` file which incorporates this
action, and if we accidentally see it then we'll mis-badger authors.

Since this thing produces a regex for incorporating into a `git grep` command,
the output of this action now includes the recommended flags. Previously the
consumer had to put these flags on, rather than receiving a hint.